moved caveat related functions to a new file
