+++ packages/rpc-methods/src/restricted/invokeSnap.ts
@@ -3,9 +3,6 @@ import {
   RestrictedMethodOptions,
   ValidPermissionSpecification,
   PermissionType,
-  RestrictedMethodCaveatSpecificationConstraint,
-  Caveat,
-  RestrictedMethodParameters,
   PermissionValidatorConstraint,
 } from '@metamask/permission-controller';
 import {
@@ -14,15 +11,8 @@ import {
   HandlerType,
   SnapRpcHookArgs,
   SnapCaveatType,
-  assertIsValidSnapId,
 } from '@metamask/snaps-utils';
-import {
-  isJsonRpcRequest,
-  Json,
-  NonEmptyArray,
-  hasProperty,
-  isObject,
-} from '@metamask/utils';
+import { isJsonRpcRequest, Json, NonEmptyArray } from '@metamask/utils';
 import { ethErrors } from 'eth-rpc-errors';
 import { nanoid } from 'nanoid';
 
@@ -53,29 +43,10 @@ type InvokeSnapSpecification = ValidPermissionSpecification<{
   validator: PermissionValidatorConstraint;
 }>;
 
-type InvokeSnapParams = {
+export type InvokeSnapParams = {
   snapId: string;
   request: Record<string, unknown>;
 };
-
-/**
- * Validates that the caveat value exists and is a non-empty object.
- *
- * @param caveat - The caveat to validate.
- * @throws If the caveat is invalid.
- */
-export function validateCaveat(caveat: Caveat<string, any>) {
-  if (!isObject(caveat.value) || Object.keys(caveat.value).length === 0) {
-    throw ethErrors.rpc.invalidParams({
-      message:
-        'Expected caveat to have a value property of a non-empty object of snap IDs.',
-    });
-  }
-  const snapIds = Object.keys(caveat.value);
-  for (const snapId of snapIds) {
-    assertIsValidSnapId(snapId);
-  }
-}
 
 /**
  * The specification builder for the `wallet_snap_*` permission.
@@ -119,32 +90,6 @@ export const invokeSnapBuilder = Object.freeze({
   specificationBuilder,
   methodHooks,
 } as const);
-
-export const InvokeSnapCaveatSpecifications: Record<
-  SnapCaveatType.SnapIds,
-  RestrictedMethodCaveatSpecificationConstraint
-> = {
-  [SnapCaveatType.SnapIds]: Object.freeze({
-    type: SnapCaveatType.SnapIds,
-    validator: (caveat) => validateCaveat(caveat),
-    decorator: (method, caveat) => {
-      return async (args) => {
-        const {
-          params,
-          context: { origin },
-        }: RestrictedMethodOptions<RestrictedMethodParameters> = args;
-        const snapIds = caveat.value;
-        const { snapId } = params as InvokeSnapParams;
-        if (!hasProperty(snapIds, snapId)) {
-          throw new Error(
-            `${origin} does not have permission to invoke ${snapId} snap.`,
-          );
-        }
-        return await method(args);
-      };
-    },
-  }),
-};
 
 /**
  * Builds the method implementation for `wallet_snap_*`.
